fix potential 2s complements type bug nightmare

+++ usaspending_api/etl/management/commands/update_total_funding_amount_sql.py
@@ -42,7 +42,7 @@ class Command(BaseCommand):
 
     SUM_TRANSACTION_TFA = """
         With sum_table as (
-        SELECT award_id, SUM(NULLIF(tn.funding_amount, '')::float) sum_val 
+        SELECT award_id, SUM(NULLIF(tn.funding_amount, '')::decimal) sum_val
         FROM transaction_normalized as tn join awards a on tn.award_id = a.id WHERE a.id = tn.award_id GROUP BY award_id
         )
         update awards as a
